The test CloudBlobContainerTests.testSASParserBug() can reproduce a concurrency bug in Utility.saxParserThreadLocal. Sometimes you need to run the test a few times, but it is reproducing fairly easily for me.
saxParserThreadLocal is a ThreadLocal<SAXParser> with a SAXParserFactory member variable.
This factory is initialized by ThreadLocal.initialValue which then continues to use
the member variable reference.  Because initialValue can be called by multiple threads,
it is possible for Thread A to be after the call to "factory.setNamespaceAware(true)" but
before the call to "return factory.newSAXParser()" when it yields to Thread B which
calls "factory = SAXParserFactory.newInstance()" and then yields to Thread A which
then calls "return factory.newSAXParser()".  This factory instance is therefore not
namespace aware since setNamespaceAware(true) has not been called.

A SAXParser that is not namespace aware will always return a list count of 0.  This was
also confirmed in a unit test, but for brevity this test was not included in this commit.