SAXParser concurrency bug fix.
DETAILS: Utility.saxParserThreadLocal is a ThreadLocal<SAXParser> with a SAXParserFactory member variable.
The factory is initialized in ThreadLocal.initialValue which then continues to use the member variable
reference.  Because initialValue can be called by multiple threads,
it is possible for Thread A to be after the call to "factory.setNamespaceAware(true)" but
before the call to "return factory.newSAXParser()" when it yields to Thread B which
calls "factory = SAXParserFactory.newInstance()" and then yields to Thread A which
then calls "return factory.newSAXParser()".  This factory instance is therefore not
namespace aware since setNamespaceAware(true) has not been called.

A SAXParser that is not namespace aware will always return a list count of 0.  This was
also confirmed in a unit test, but for brevity this test was not included in this commit.

TESTS: The test testSASParserConcurrency was added to validate that the SAXParser returned by Utility.getSAXParser
is correctly configured, even when called under a highly concurrent load.